feat(Messaging): Moved messaging related methods to separate module
